Add files via upload
It is an academic project that focuses on analyzing the relationship between demographic and socio-economic factors and academic performance using SPSS software. The project includes data cleaning to ensure the data is accurate and complete, as well as handling missing values. Descriptive statistics are used to summarize the data, and inferential statistical tests such as ANOVA and chi-square are used to determine the significance of any relationships found. Linear and logistic regression are also used to model and predict academic performance based on demographic and socio-economic factors. 
The goal of the project is to understand how these factors may impact student success and identify areas for improvement.