Data collection and pre-processing using Python
This project involved conducting research on greenhouse gas emissions on a global scale, which required a significant amount of data collection and manipulation. At the beginning of the project, the data was collected from a variety of sources and was quite messy. However, as the project progressed, I was able to develop and implement a systematic process for collecting and formatting data for each country.
The data collection process involved:
•	Identifying and sourcing relevant data from various online databases and government websites
•	Cleaning and pre-processing the data using Python and libraries such as Pandas and NumPy
The data formatting process involved:
•	Standardizing the format of the data to ensure consistency across all countries
•	Translating data into English where necessary
•	Handling missing and duplicate data
•	Creating calculated fields and measures to facilitate analysis
The end result of this process was a clean and consistent dataset that could be easily analyzed to answer the research questions. This process proved to be efficient and effective in handling the large volume of data needed for this international research project.
This required ability to work with large and complex datasets, implement efficient data manipulation and cleaning techniques, and effectively use tools such as Python and SQL in a data science pipeline.